requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.11
Django==3.2
django-common-helpers==0.9.2
django-cron==0.5.1
django-extensions==3.1.5
django-filter==21.1
django-restframework==0.0.1
djangorestframework==3.13.1
idna==3.3
psycopg2==2.9.3
python-dotenv==0.19.2
pytz==2021.3
requests==2.27.1
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.8
